+++ http/users.go
@@ -54,11 +54,12 @@ func userUpdateRevokesSessions(which []string) bool {
 	return false
 }
 
-func getUser(_ http.ResponseWriter, r *http.Request) (*modifyUserRequest, error) {
+func getUser(w http.ResponseWriter, r *http.Request) (*modifyUserRequest, error) {
 	if r.Body == nil {
 		return nil, fberrors.ErrEmptyRequest
 	}
 
+	capJSONBody(w, r)
 	req := &modifyUserRequest{}
 	err := json.NewDecoder(r.Body).Decode(req)
 	if err != nil {
@@ -122,7 +123,7 @@ var userGetHandler = withSelfOrAdmin(func(w http.ResponseWriter, r *http.Request
 	return renderJSON(w, r, u)
 })
 
-var userDeleteHandler = withSelfOrAdmin(func(_ http.ResponseWriter, r *http.Request, d *data) (int, error) {
+var userDeleteHandler = withSelfOrAdmin(func(w http.ResponseWriter, r *http.Request, d *data) (int, error) {
 	if r.Body == nil {
 		return http.StatusBadRequest, fberrors.ErrEmptyRequest
 	}
@@ -131,6 +132,7 @@ var userDeleteHandler = withSelfOrAdmin(func(_ http.ResponseWriter, r *http.Requ
 		CurrentPassword string `json:"current_password"`
 	}
 
+	capJSONBody(w, r)
 	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
 		return http.StatusBadRequest, err
 	}